refactor prompt html and jsdocs
